strategies/sniping_strategy.rs
use crate::models::market_conditions::MarketConditions;
use crate::models::sniping_opportunity::SnipingOpportunity;
use crate::strategies::strategy::Strategy;
use crate::utils::solana::{analyze_transaction, calculate_profit};
use async_trait::async_trait;
use solana_sdk::pubkey::Pubkey;
use std::collections::HashMap;

pub struct SnipingStrategy {
    pub rpc_client: solana_client::rpc_client::RpcClient,
    pub target_tokens: Vec<Pubkey>,
    pub max_price: f64,
    pub min_liquidity: f64,
}

impl SnipingStrategy {
    pub fn new(rpc_client: solana_client::rpc_client::RpcClient) -> Self {
        SnipingStrategy {
            rpc_client,
            target_tokens: Vec::new(),
            max_price: 0.0,
            min_liquidity: 0.0,
        }
    }

    pub fn add_target_token(&mut self, token_mint: Pubkey) {
        self.target_tokens.push(token_mint);
    }

    pub fn set_max_price(&mut self, max_price: f64) {
        self.max_price = max_price;
    }

    pub fn set_min_liquidity(&mut self, min_liquidity: f64) {
        self.min_liquidity = min_liquidity;
    }
}

#[async_trait]
impl Strategy for SnipingStrategy {
    fn update(&mut self, market_conditions: &MarketConditions) {
        self.max_price = market_conditions.max_price;
        self.min_liquidity = market_conditions.min_liquidity;
    }

    async fn find_opportunities(&self, target_accounts: &HashMap<Pubkey, crate::AccountInfo>) -> Vec<SnipingOpportunity> {
        let mut opportunities = Vec::new();
        
        for (token_mint, account_info) in target_accounts {
            if self.target_tokens.contains(token_mint) {
                let token_balance = account_info.token_balance;
                let token_price = account_info.token_price;
                let token_liquidity = self.get_token_liquidity(token_mint).await;
                
                if token_price <= self.max_price && token_liquidity >= self.min_liquidity {
                    let opportunity = SnipingOpportunity {
                        target_account: *token_mint,
                        token_mint: *token_mint,
                        expected_price: token_price,
                        token_balance,
                    };
                    opportunities.push(opportunity);
                }
            }
        }

        opportunities
    }
}

impl SnipingStrategy {
    async fn get_token_liquidity(&self, token_mint: &Pubkey) -> f64 {
        let liquidity = 0.0;
        liquidity
    }
    
    async fn execute_sniping(&self, opportunity: &SnipingOpportunity) -> Option<String> {
        let transaction = self.build_sniping_transaction(opportunity);
        let signature = self.rpc_client.send_transaction(&transaction).await;
        
        if let Ok(signature) = signature {
            let transaction_info = self.rpc_client.get_transaction(&signature).await.unwrap();
            let profit = calculate_profit(&transaction_info).unwrap();
            println!("Sniping transaction executed. Profit: {}", profit);
            Some(signature.to_string())
        } else {
            println!("Failed to execute sniping transaction");
            None
        }
    }
    
    fn build_sniping_transaction(&self, opportunity: &SnipingOpportunity) -> solana_sdk::transaction::Transaction {
        let transaction = solana_sdk::transaction::Transaction::new_with_payer(
            &[],
            Some(&self.rpc_client.payer_pubkey()),
        );
        transaction
    }
}